refactor: 컴포넌트를 개별파일로 분리한다
