+++ src/components/Button.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Button({ children }) {
+  return (
+    <button type="button">
+      {children}
+    </button>
+  );
+}
