+++ src/components/Counter.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function Counter({ count, onClick }) {
+  return (
+    <button type="button" onClick={onClick}>
+      Click me!
+      (
+      {count}
+      )
+    </button>
+  );
+}
